fix overflow and update python package

+++ src/rematch/evaluation/algorithm/stream_algorithm.cpp
@@ -150,7 +150,7 @@ void StreamAlgorithm::update_sets(
 
 void StreamAlgorithm::update_output_nodes(ExtendedDetVAState*& next_state,
                                           ECSNode*& next_node) {
-  if (next_state->phase < (int64_t)pos_i_) {
+  if (next_state->phase < pos_i_) {
     next_state->set_phase(pos_i_);
 
     next_state->set_node(next_node);
